Initial work on implementation of server registration and activation
Implement the IntelliSense Server registration in environment variables and activation/deactivation via an Excel macro. Still outstanding is the where the active server retrieves registration info from loaded add-ins (or elsewhere).

This version is still very unstable, especially when the Excel 2013 startup page is loaded, and on shutdown. In-sheet function descriptions are also not initially displayed.